# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: PayGate Payment Platform API
  version: 1.0.0
  description: |
    API contract for the current MVP implementation of the PayGate multi-tenant
    payment middleware.

    Coverage in this spec matches the endpoints that are already live in the
    backend codebase today:
    - dashboard auth and MFA
    - store management
    - store API token management
    - store-facing transaction API
    - Midtrans inbound webhook
    - webhook delivery inspection and resend
    - dashboard operational alert destinations
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/dashboard/auth/mfa/recovery/regenerate:
    post:
      tags: [Dashboard Auth]
      summary: Regenerate MFA recovery codes
      operationId: regenerateRecoveryCodes
      security:
        - DashboardBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MFACodeRequest"
      responses:
        "200":
          description: New recovery codes generated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RecoveryCodesResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/MFAInvalidCodeError"
        "409":
          $ref: "#/components/responses/MFASetupRequiredError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/dashboard/me:
    get:
      tags: [Dashboard Auth]
      summary: Get current dashboard user
      operationId: getDashboardMe
      security:
        - DashboardBearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/dashboard/stores:
    get:
      tags: [Dashboard Stores]
      summary: List stores for current user
      operationId: listStores
      security:
        - DashboardBearerAuth: []
      responses:
        "200":
          description: Store list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/StoreListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/DashboardAccessError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Dashboard Stores]
      summary: Create store
      operationId: createStore
      security:
        - DashboardBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoreRequest"
            example:
              name: Demo Store
              slug: demo-store
              domain: shop.example.com
              default_callback_url: https://shop.example.com/api/payment/callback
      responses:
        "201":
          description: Store created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/StoreWithSecret"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/DashboardAccessError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/dashboard/alert-endpoints:
    get:
      tags: [Dashboard Alerts]
      summary: List operational alert endpoints
      operationId: listAlertEndpoints
      security:
        - DashboardBearerAuth: []
      responses:
        "200":
          description: Operational alert destinations owned by the current dashboard user
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AlertEndpointListResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/DashboardAccessError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Dashboard Alerts]
      summary: Create operational alert endpoint
      operationId: createAlertEndpoint
      security:
        - DashboardBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAlertEndpointRequest"
            example:
              name: Slack Ops
              channel: slack_webhook
              destination_url: https://hooks.slack.com/services/T000/B000/XYZ
              events:
                - webhook.failed_permanently
              status: active
      responses:
        "201":
          description: Alert endpoint created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AlertEndpoint"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/DashboardAccessError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/dashboard/alert-endpoints/{endpoint_id}:
    patch:
      tags: [Dashboard Alerts]
      summary: Update operational alert endpoint
      operationId: updateAlertEndpoint
      security:
        - DashboardBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AlertEndpointID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAlertEndpointRequest"
            example:
              name: Slack Ops Updated
              channel: webhook
              destination_url: https://ops.example.com/hooks/paygate
              events:
                - webhook.failed_permanently
              status: active
              clear_auth_token: false
      responses:
        "200":
          description: Alert endpoint updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AlertEndpoint"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/DashboardAccessError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Dashboard Alerts]
      summary: Delete operational alert endpoint
      operationId: deleteAlertEndpoint
      security:
        - DashboardBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AlertEndpointID"
      responses:
        "204":
          description: Alert endpoint deleted
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/DashboardAccessError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/dashboard/alert-endpoints/{endpoint_id}/test:
    post:
      tags: [Dashboard Alerts]
      summary: Send test alert to an operational endpoint
      operationId: sendTestAlertEndpoint
      security:
        - DashboardBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AlertEndpointID"
      responses:
        "202":
          description: Test alert accepted and queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AlertEndpointTestDispatch"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/DashboardAccessError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/transactions/charge:
    post:
      tags: [Store API]
      summary: Create transaction and proxy charge to Midtrans
      description: |
        Store-facing endpoint ini hanya menerima request server-to-server. Request browser yang membawa `Origin` atau `Sec-Fetch-*` akan diblokir dengan `403 BROWSER_REQUEST_BLOCKED`.
      operationId: chargeTransaction
      security:
        - StoreBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChargeRequest"
            example:
              order_id: INV-2026-0001
              amount: 150000
              currency: IDR
              payment_type: bank_transfer
              bank: bca
              customer:
                name: Budi
                email: budi@example.com
                phone: "08123456789"
              items:
                - id: SKU-001
                  name: Produk A
                  price: 150000
                  quantity: 1
              callback_url: https://tokouser.com/api/payment/callback
              metadata:
                cart_id: cart_123
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ChargeResult"
              example:
                success: true
                data:
                  transaction_id: 7dd41ec0-1c48-492a-b2df-8be28b4115b8
                  order_id: INV-2026-0001
                  platform_order_id: store123_INV-2026-0001
                  status: pending
                  payment_type: bank_transfer
                  amount: 150000
                  midtrans:
                    transaction_id: midtrans_transaction_id
                    va_numbers:
                      - bank: bca
                        va_number: "1234567890123456"
                    transaction_status: pending
                    fraud_status: accept
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/StoreUnauthorizedError"
        "403":
          $ref: "#/components/responses/StoreForbiddenError"
        "409":
          $ref: "#/components/responses/TransactionConflictError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
        "502":
          $ref: "#/components/responses/MidtransError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/transactions/{order_id}:
    get:
      tags: [Store API]
      summary: Get transaction by store order ID
      operationId: getTransactionByOrderID
      security:
        - StoreBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/OrderID"
      responses:
        "200":
          description: Transaction detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Transaction"
              example:
                success: true
                data:
                  id: 7dd41ec0-1c48-492a-b2df-8be28b4115b8
                  order_id: INV-2026-0001
                  platform_order_id: store123_INV-2026-0001
                  midtrans_transaction_id: midtrans_transaction_id
                  payment_type: bank_transfer
                  gross_amount: 150000
                  currency: IDR
                  status: paid
                  fraud_status: accept
                  metadata:
                    cart_id: cart_123
                  created_at: "2026-05-02T10:25:12Z"
                  updated_at: "2026-05-02T10:31:00Z"
                  paid_at: "2026-05-02T10:30:00Z"
        "401":
          $ref: "#/components/responses/StoreUnauthorizedError"
        "403":
          $ref: "#/components/responses/StoreForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/audit-logs:
    get:
      tags: [Store API]
      summary: List store audit logs
      description: |
        Endpoint audit log store hanya boleh dipanggil server-to-server memakai store API token. Request browser yang membawa `Origin` atau `Sec-Fetch-*` akan diblokir dengan `403 BROWSER_REQUEST_BLOCKED`.
      operationId: listStoreAuditLogs
      security:
        - StoreBearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/AuditLogDirection"
        - $ref: "#/components/parameters/AuditLogRequestID"
        - $ref: "#/components/parameters/AuditLogOrderID"
        - $ref: "#/components/parameters/AuditLogEndpoint"
        - $ref: "#/components/parameters/AuditLogStatusCode"
        - $ref: "#/components/parameters/AuditLogCreatedFrom"
        - $ref: "#/components/parameters/AuditLogCreatedTo"
        - $ref: "#/components/parameters/Query"
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuditLogListResponse"
              example:
                success: true
                data:
                  logs:
                    - id: 2f18a945-0f89-4713-958e-7dc238cb4d9d
                      request_id: req_20260502_001
                      actor_type: store_api
                      actor_id: 9ab66fb6-cfbd-44ba-8df9-499c53e55d12
                      direction: inbound
                      method: POST
                      url: /v1/transactions/charge
                      status_code: 201
                      request_body:
                        order_id: INV-2026-0001
                        amount: 150000
                      response_body:
                        transaction_id: 7dd41ec0-1c48-492a-b2df-8be28b4115b8
                        status: pending
                      error_message: null
                      duration_ms: 82
                      created_at: "2026-05-02T10:31:00Z"
                  meta:
                    total: 17
                    limit: 10
                    offset: 0
                    has_next: true
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/StoreUnauthorizedError"
        "403":
          $ref: "#/components/responses/StoreForbiddenError"
        "429":
          $ref: "#/components/responses/RateLimitedError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/webhooks/midtrans:
    post:
      tags: [Midtrans Webhook]
      summary: Receive Midtrans webhook notification
      operationId: processMidtransWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MidtransWebhookNotification"
            example:
              transaction_time: "2026-05-02 10:25:11"
              transaction_status: settlement
              transaction_id: midtrans_transaction_id
              status_message: midtrans payment notification
              status_code: "200"
              signature_key: 2b4f7d...
              payment_type: bank_transfer
              order_id: store123_INV-2026-0001
              merchant_id: G002540291
              gross_amount: "150000.00"
              fraud_status: accept
              settlement_time: "2026-05-02 10:30:00"
      responses:
        "200":
          description: Webhook recorded and processed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MidtransWebhookResult"
              example:
                success: true
                data:
                  webhook_id: 20d73a69-f3e2-44ea-8fca-d7ba17fcd413
                  transaction_id: 7dd41ec0-1c48-492a-b2df-8be28b4115b8
                  platform_order_id: store123_INV-2026-0001
                  transaction_found: true
                  signature_valid: true
                  internal_status: paid
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/WebhookSignatureInvalidError"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    DashboardBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    StoreBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Store API Token
  parameters:
...
    AlertEndpointID:
      name: endpoint_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    OrderID:
      name: order_id
      in: path
      required: true
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
    Query:
      name: query
      in: query
      required: false
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid payload.
              request_id: req_abc123
              details: {}
    UnauthorizedError:
      description: Dashboard authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Invalid or expired access token.
              request_id: req_abc123
              details: {}
    StoreUnauthorizedError:
      description: Store API authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Invalid store API token.
              request_id: req_abc123
              details: {}
...
    DashboardAccessError:
      description: Dashboard access blocked by MFA policy
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            setup_required:
              value:
                success: false
                error:
                  code: MFA_SETUP_REQUIRED
                  message: MFA setup is required before dashboard access.
                  request_id: req_abc123
                  details: {}
            verification_required:
              value:
                success: false
                error:
                  code: MFA_VERIFICATION_REQUIRED
                  message: MFA verification is required before dashboard access.
                  request_id: req_abc123
                  details: {}
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found.
              request_id: req_abc123
              details: {}
    ConflictError:
      description: Unique or state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: CONFLICT
              message: Resource already exists.
              request_id: req_abc123
              details: {}
...
    MFASetupRequiredError:
      description: MFA has not been enabled yet
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: MFA_SETUP_REQUIRED
              message: MFA is not enabled yet.
              request_id: req_abc123
              details: {}
...
    MFAInvalidCodeError:
      description: MFA code or recovery code is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: MFA_CODE_INVALID
              message: Invalid MFA or recovery code.
              request_id: req_abc123
              details: {}
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: Internal server error.
              request_id: req_abc123
              details: {}
  schemas:
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
          const: true
        data: {}
...
    MFACodeRequest:
      type: object
      required: [code]
      properties:
        code:
          type: string
          description: TOTP code or recovery code depending on endpoint and state
...
    RecoveryCodesResponse:
      type: object
      required: [recovery_codes]
      properties:
        recovery_codes:
          type: array
          items:
            type: string
...
    MeResponse:
      type: object
      required: [user, mfa]
      properties:
        user:
          $ref: "#/components/schemas/User"
        mfa:
          $ref: "#/components/schemas/MFAState"
    AlertEndpoint:
      type: object
      required:
        [id, user_id, name, channel, destination_url, events, status, has_auth_token, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        channel:
          type: string
          enum: [webhook, slack_webhook, discord_webhook]
        destination_url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum: [webhook.failed_permanently]
        status:
          type: string
          enum: [active, inactive]
        has_auth_token:
          type: boolean
        last_tested_at:
          type: string
          format: date-time
          nullable: true
        last_triggered_at:
          type: string
          format: date-time
          nullable: true
        last_success_at:
          type: string
          format: date-time
          nullable: true
        last_error:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AlertEndpointListResponse:
      type: object
      required: [endpoints]
      properties:
        endpoints:
          type: array
          items:
            $ref: "#/components/schemas/AlertEndpoint"
    CreateAlertEndpointRequest:
      type: object
      required: [name, channel, destination_url]
      properties:
        name:
          type: string
        channel:
          type: string
          enum: [webhook, slack_webhook, discord_webhook]
        destination_url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum: [webhook.failed_permanently]
        status:
          type: string
          enum: [active, inactive]
        auth_token:
          type: string
          writeOnly: true
    UpdateAlertEndpointRequest:
      type: object
      properties:
        name:
          type: string
        channel:
          type: string
          enum: [webhook, slack_webhook, discord_webhook]
        destination_url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum: [webhook.failed_permanently]
        status:
          type: string
          enum: [active, inactive]
        auth_token:
          type: string
          writeOnly: true
        clear_auth_token:
          type: boolean
    AlertEndpointTestDispatch:
      type: object
      required: [alert_delivery_id, status]
      properties:
        alert_delivery_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending]
...
    StoreWithSecret:
      allOf:
        - $ref: "#/components/schemas/Store"
        - type: object
          required: [webhook_secret]
          properties:
            webhook_secret:
              type: string
...
    StoreListResponse:
      type: object
      required: [stores]
      properties:
        stores:
          type: array
          items:
            $ref: "#/components/schemas/Store"
    CreateStoreRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        slug:
          type: string
        domain:
          type: string
        default_callback_url:
          type: string
          format: uri
...
    ChargeRequest:
      type: object
      required:
        [order_id, amount, currency, payment_type, bank, customer, items]
      properties:
        order_id:
          type: string
        amount:
          type: integer
          minimum: 1
        currency:
          type: string
          example: IDR
        payment_type:
          type: string
          example: bank_transfer
        bank:
          type: string
          enum: [bca, bni, bri, cimb, permata, mandiri]
        customer:
          $ref: "#/components/schemas/Customer"
        items:
          type: array
          items:
            $ref: "#/components/schemas/Item"
        callback_url:
          type: string
          format: uri
        metadata:
          type: object
          additionalProperties: true
...